Make cv2 "soft" dependency, create oak_outputs.xout subpackage
