+++ depthai_sdk/src/depthai_sdk/__init__.py
@@ -1,15 +1,4 @@
 from depthai_sdk.args_parser import ArgsParser
-from depthai_sdk.classes import *
-from depthai_sdk.classes.packets import *
-from depthai_sdk.components import *
 from depthai_sdk.components.nn_helper import ResizeMode
-from depthai_sdk.fps import *
 from depthai_sdk.oak_camera import OakCamera
 from depthai_sdk.oak_device import OakDevice
-from depthai_sdk.oak_outputs.fps import FPS
-from depthai_sdk.previews import *
-from depthai_sdk.record import *
-from depthai_sdk.replay import *
-from depthai_sdk.utils import *
-from depthai_sdk.visualize import *
-from depthai_sdk.visualize.visualizer_helper import VisualizerHelper
