+++ depthai_sdk/src/depthai_sdk/classes/output_config.py
@@ -4,13 +4,14 @@ from typing import Optional, Callable, List
 
 import depthai as dai
 
-from depthai_sdk import FramePacket
+from depthai_sdk.classes import FramePacket
 from depthai_sdk.oak_outputs.syncing import SequenceNumSync
-from depthai_sdk.oak_outputs.xout import XoutFrames, XoutDepth
-from depthai_sdk.oak_outputs.xout_base import XoutBase
+from depthai_sdk.oak_outputs.xout.xout_depth import XoutDepth
+from depthai_sdk.oak_outputs.xout.xout_frames import XoutFrames
+from depthai_sdk.oak_outputs.xout.xout_base import XoutBase
 from depthai_sdk.record import Record
 from depthai_sdk.recorders.video_recorder import VideoRecorder
-from depthai_sdk.visualize import Visualizer
+from depthai_sdk.visualize.visualizer import Visualizer
 
 
 class BaseConfig:
