+++ depthai_sdk/src/depthai_sdk/components/component.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 import depthai as dai
 
-from depthai_sdk.oak_outputs.xout_base import XoutBase, ReplayStream
+from depthai_sdk.oak_outputs.xout.xout_base import XoutBase, ReplayStream
 
 
 class Component(ABC):
