--- depthai_sdk/src/depthai_sdk/visualize/__init__.py
@@ -1,4 +1,0 @@
-from .configs import BboxStyle, TextPosition
-from .configs import VisConfig, DetectionConfig, TextConfig, TrackingConfig
-from .objects import GenericObject, VisDetections, VisText, VisLine, VisTrail
-from .visualizer import Visualizer
